// src/lib/schema/format-enum.ts
export enum EFormat {
  DATE = 'date',
  DATE_TIME = 'date-time',
  DURATION = 'duration',
  EMAIL = 'email',
  HOSTNAME = 'hostname',
  IDN_EMAIL = 'idn-email',
  IDN_HOSTNAME = 'idn-hostname',
  IRI = 'iri',
  IRI_REFERENCE = 'iri-reference',
  IPV4 = 'ipv4',
  IPV6 = 'ipv6',
  JSON_POINTER = 'json-pointer',
  REGEX = 'regex',
  RELATIVE_JSON_POINTER = 'relative-json-pointer',
  TIME = 'time',
  URI = 'uri',
  URI_REFERENCE = 'uri-reference',
  URI_TEMPLATE = 'uri-template',
  UUID = 'uuid',
}