// src/lib/schema/restriction-enum.ts
export enum ERestriction {
  CONST = 'const',
  ENUM = 'enum',
  EXCLUSIVE_MAXIMUM = 'exclusiveMaximum',
  EXCLUSIVE_MINIMUM = 'exclusiveMinimum',
  FORMAT = 'format',
  MAX_ITEMS = 'maxItems',
  MAX_LENGTH = 'maxLength',
  MAXIMUM = 'maximum',
  MIN_ITEMS = 'minItems',
  MIN_LENGTH = 'minLength',
  MINIMUM = 'minimum',
  MULTIPLE_OF = 'multipleOf',
  PATTERN = 'pattern',
  UNIQUE_ITEMS = 'uniqueItems'
}